+++ lib/src/core/utils/daemon_manager/bloc/daemon_cubit.dart
@@ -116,7 +116,7 @@ class DaemonCubit extends Cubit<DaemonState> {
       // bypass password-less wallet cli and init node
       if (cliCommand.command == './pactus-daemon' &&
           cliCommand.arguments.first == 'init' &&
-          !cliCommand.arguments.contains('--password')) {
+          !cliCommand.arguments.contains('password')) {
         // Writing password interactively
         process.stdin.writeln();
 
@@ -127,8 +127,8 @@ class DaemonCubit extends Cubit<DaemonState> {
       }
 
       if (cliCommand.command == './pactus-wallet' &&
-          cliCommand.arguments.first == 'password') {
-        final pass = cliCommand.arguments.last;
+          cliCommand.arguments[0] == 'password') {
+        final pass = cliCommand.arguments[2];
         // Writing password interactively
         process.stdin.writeln(pass);